requirements.txt
SQLAlchemy==2.0.30
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pytest==8.2.1
httpx==0.27.0
asyncpg==0.29.0
postgres==4.0
pydantic-settings==2.2.1
FastAPI-SQLAlchemy==0.2.1
